Use crypto/subtle.XORBytes
The function XORBytes has been included in the Go standard library
since Go version 1.20.  This allows us to remove a bunch of
assembler code, and instead rely on the stdlib.

There are only three versions remaining:

  - xor_generic, which simply calls the stdlib function;
  - xor_old, which uses unsafe Go and is used with Go 1.19;
  - xor_arm, which is written in assembly and is used on
    32-bit ARM, since the stdlib doesn't implement a fast
    version of XORBytes on that architecture.
